pyproject.toml
[tool.poetry]
name = "markdown"
version = "0.1.0"
description = ""
authors = ["Isabelle-C <isabellechen2023@u.northwestern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pypandoc = "^1.13"
click = "^8.1.7"
pypdf = "^4.2.0"
pymupdf = "^1.24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
